requirements.txt
asgiref==3.6.0
autopep8==2.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptocode==0.1
cryptography==38.0.4
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
google-auth==2.15.0
google-auth-oauthlib==0.8.0
idna==3.4
importlib-metadata==5.2.0
mailchimp-transactional==1.0.50
Markdown==3.4.1
model-bakery==1.9.0
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.1
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
pip==22.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.16.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
types-pytz==2022.7.0.0
types-requests==2.28.11.7
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.13
wheel==0.38.4
zipp==3.11.0
